Move ArenaParticipant model to Models
