+++ ArenaService/ArenaService/ArenaParticipantType.cs
@@ -1,4 +1,5 @@
 using GraphQL.Types;
+using ArenaService.Models;
 
 namespace ArenaService;
