+++ ArenaService/ArenaService/ArenaWorker.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using ArenaService.Models;
 
 namespace ArenaService;
